refactor(redux): move generic redux types and functions into helper file
