+++ src/domain/wallet/store/store.ts
@@ -26,22 +26,3 @@ export type ReduxStore = Store<AppState> & {
   readonly dispatch: ThunkDispatch<AppState, Dependencies, Action>
 }
 export type ThunkResult<R> = ThunkAction<R, AppState, Dependencies, Action>
-
-export interface ActionWithPayload<T extends string, P> extends Action<T> {
-  readonly payload: P
-}
-
-export function createAction<T extends string>(type: T): Action<T>
-export function createAction<T extends string, P>(type: T, payload?: P): ActionWithPayload<T, P>
-
-export function createAction<T extends string, P>(
-  type: T,
-  payload?: P
-): Action<T> | ActionWithPayload<T, P> {
-  return payload === undefined ? { type } : { type, payload }
-}
-
-type FunctionType = (...args: readonly never[]) => unknown
-type ActionCreatorsMapObject = Record<string, FunctionType>
-
-export type ActionsUnion<A extends ActionCreatorsMapObject> = ReturnType<A[keyof A]>
